[iqm] Rework iqm joints and animations
The main purpose of reworking the joints was to rearrange the data so
the rotation quaternion was aligned for vec4, but as this needed
reworking the rest of the loader, I thought I'd get dual quaternions
(ish) working. They do... for the most part: shear's not representable
(nor desired, usually), and Mr Fixit doesn't like interpolated
quaternions: the quaternions flip signs in the animation resulting in
some very strange deformations, and he was modeled with matrix
interpolation in mind.

Also, there are linker issues caused by the animation code being in the
renderer (where it most definitely does not belong).