[sw] Fix a pile of unclipped alias model issues
My tests that showed the sw renderer was working turned out to be
incomplete because only the view model was visible thus only the clipped
path was taken. The unclipped path proved to be a little trickier to
debug but in the end was due to global variables not getting set
properly (oh, how I hate global variables).