[qfbsp] Use better duplicate plane check
Testing whether the points are within epsilon of a plane doesn't work
too well when they describe a plane at a very different angle (though
the plane being that small is a problem in itself). This seems to have
taken care of the bogus coordinates, but it has uncovered all sorts of
other problems.