[qfcc] Branch struct and array types for blocks
SPIR-V buffer block members need to be fully decorated with offsets
(recursively), which means that any members with struct or array types
need their own types in order to support having different offset and
stride (arrays) decorations. This breaks struct assignment at the
language level, but that is intentional as I want to deal with the
member copying needed to implement struct assignment before emitting
code. It also removes the need for some of the complication in
target_spirv.c, but I'll deal with that later.

The recursively built tag for nested structs might be a bit too
aggressive, but that too can wait until I have a better idea of what's
needed.